Now supports monocle and ignores floating window
Floating is being ignored to accommodate for scratchpads (ie., dropdown
terminal). Of course, this is not for everyone, but this works the best
for me who uses bunch of dropdown terminals.

Monocle is one of those features that I also use quite often which this
script did not support initially, so I added it.